Feature/spawning and moving enemies (#28)
* Refactor texture scaling vector construction

* Implement enemy node

+++ BasicShmup/Src/Entities/Projectiles/ProjectileConfiguration.cs
@@ -22,7 +22,7 @@ public partial class ProjectileConfiguration : Resource, IConfiguration, IProjec
     public Texture2D Texture { get; set; } = ResourceLoader.Load<Texture2D>("res://Resources/ErrorTexture.png");
 
     [Export]
-    public int TextureScaling { get; set; } = 1;
+    public float TextureScaling { get; set; } = 1;
 
     public void Register(IServiceCollection serviceCollection)
     {
